+++ src/kernels/compute_inner_integral_numerically.cu
@@ -31,16 +31,16 @@ void compute_G_numeric(
         // fiberfloat TGF1;
         // fiberfloat TGF2;
 
-        G[quadrature_index_i + 0 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0;
-        G[quadrature_index_i + 1 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0;
-        G[quadrature_index_i + 2 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0;
-        G[quadrature_index_i + 3 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0;
-        G[quadrature_index_i + 4 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0;
-        G[quadrature_index_i + 5 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0;
+        G[quadrature_index_i + 0 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0f;
+        G[quadrature_index_i + 1 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0f;
+        G[quadrature_index_i + 2 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0f;
+        G[quadrature_index_i + 3 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0f;
+        G[quadrature_index_i + 4 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0f;
+        G[quadrature_index_i + 5 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0f;
 
-        GF[quadrature_index_i + 0 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0;
-        GF[quadrature_index_i + 1 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0;
-        GF[quadrature_index_i + 2 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0;
+        GF[quadrature_index_i + 0 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0f;
+        GF[quadrature_index_i + 1 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0f;
+        GF[quadrature_index_i + 2 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0f;
 
         fiberfloat4 position_on_fiber_i;
         position_on_fiber_i.x = position_i.x + quadrature_points[quadrature_index_i] * orientation_i.x;
@@ -60,35 +60,35 @@ void compute_G_numeric(
             difference.y = position_on_fiber_i.y - position_on_fiber_j.y;
             difference.z = position_on_fiber_i.z - position_on_fiber_j.z;
 
-            const fiberfloat distance = sqrt(difference.x * difference.x + difference.y * difference.y + difference.z * difference.z);
+            const fiberfloat distance = sqrtf(difference.x * difference.x + difference.y * difference.y + difference.z * difference.z);
 
             // equation 10
             // Note:    The outer product of a vector with itself is always a symmetric matrix
             //          so to save computation we only compute the upper triangle.
             // TODO calculation can be optimized (i.e. not dividing by distance, simpifing etc.)
-            const fiberfloat K11 = 1.0 / distance
-                                   + (1.0 / distance) * (difference.x / distance) * (difference.x / distance)
-                                   + 2.0 * SLENDERNESS * SLENDERNESS * ((1.0 / (distance * distance * distance))
-                                           - (3.0 / (distance * distance * distance)) * ((difference.x / distance) * (difference.x / distance)));
-            const fiberfloat K22 = 1.0 / distance
-                                   + (1.0 / distance) * (difference.y / distance) * (difference.y / distance)
-                                   + 2.0 * SLENDERNESS * SLENDERNESS * ((1.0 / (distance * distance * distance))
-                                           - (3.0 / (distance * distance * distance)) * ((difference.y / distance) * (difference.y / distance)));
-            const fiberfloat K33 = 1.0 / distance
-                                   + (1.0 / distance) * (difference.z / distance) * (difference.z / distance)
-                                   + 2.0 * SLENDERNESS * SLENDERNESS * ((1.0 / (distance * distance * distance))
-                                           - (3.0 / (distance * distance * distance)) * ((difference.z / distance) * (difference.z / distance)));
-            const fiberfloat K12 = (1.0 / distance) * (difference.x / distance) * (difference.y / distance)
-                                   + 2.0 * SLENDERNESS * SLENDERNESS
-                                   * (-3.0 / (distance * distance * distance)) * (difference.x / distance) * (difference.y / distance);
+            const fiberfloat K11 = 1.0f / distance
+                                   + (1.0f / distance) * (difference.x / distance) * (difference.x / distance)
+                                   + 2.0f * SLENDERNESS * SLENDERNESS * ((1.0f / (distance * distance * distance))
+                                           - (3.0f / (distance * distance * distance)) * ((difference.x / distance) * (difference.x / distance)));
+            const fiberfloat K22 = 1.0f / distance
+                                   + (1.0f / distance) * (difference.y / distance) * (difference.y / distance)
+                                   + 2.0f * SLENDERNESS * SLENDERNESS * ((1.0f / (distance * distance * distance))
+                                           - (3.0f / (distance * distance * distance)) * ((difference.y / distance) * (difference.y / distance)));
+            const fiberfloat K33 = 1.0f / distance
+                                   + (1.0f / distance) * (difference.z / distance) * (difference.z / distance)
+                                   + 2.0f * SLENDERNESS * SLENDERNESS * ((1.0f / (distance * distance * distance))
+                                           - (3.0f / (distance * distance * distance)) * ((difference.z / distance) * (difference.z / distance)));
+            const fiberfloat K12 = (1.0f / distance) * (difference.x / distance) * (difference.y / distance)
+                                   + 2.0f * SLENDERNESS * SLENDERNESS
+                                   * (-3.0f / (distance * distance * distance)) * (difference.x / distance) * (difference.y / distance);
 
-            const fiberfloat K13 = (1.0 / distance) * (difference.x / distance) * (difference.z / distance)
-                                   + 2.0 * SLENDERNESS * SLENDERNESS
-                                   * (-3.0 / (distance * distance * distance)) * (difference.x / distance) * (difference.z / distance);
+            const fiberfloat K13 = (1.0f / distance) * (difference.x / distance) * (difference.z / distance)
+                                   + 2.0f * SLENDERNESS * SLENDERNESS
+                                   * (-3.0f / (distance * distance * distance)) * (difference.x / distance) * (difference.z / distance);
 
-            const fiberfloat K23 = (1.0 / distance) * (difference.y / distance) * (difference.z / distance)
-                                   + 2.0 * SLENDERNESS * SLENDERNESS
-                                   * (-3.0 / (distance * distance * distance)) * (difference.y / distance) * (difference.z / distance);
+            const fiberfloat K23 = (1.0f / distance) * (difference.y / distance) * (difference.z / distance)
+                                   + 2.0f * SLENDERNESS * SLENDERNESS
+                                   * (-3.0f / (distance * distance * distance)) * (difference.y / distance) * (difference.z / distance);
 
             const fiberfloat quadrature_weight = quadrature_weights[quadrature_index_j];
             const fiberfloat legendre_polynomial = legendre_polynomials[quadrature_index_j + force_index * TOTAL_NUMBER_OF_QUADRATURE_POINTS];
@@ -101,9 +101,9 @@ void compute_G_numeric(
             //     GF[quadrature_index_i + 1 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = quadrature_weight * (K12 * external_force.x + K22 * external_force.y + K23 * external_force.z);
             //     GF[quadrature_index_i + 2 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = quadrature_weight * (K13 * external_force.x + K23 * external_force.y + K33 * external_force.z);
 
-            //     CGF0 = 0.0;
-            //     CGF1 = 0.0;
-            //     CGF2 = 0.0;
+            //     CGF0 = 0.0f;
+            //     CGF1 = 0.0f;
+            //     CGF2 = 0.0f;
             // } else {
             //     YGF0 = quadrature_weight * (K11 * external_force.x + K12 * external_force.y + K13 * external_force.z) - CGF0;
             //     YGF1 = quadrature_weight * (K12 * external_force.x + K22 * external_force.y + K23 * external_force.z) - CGF1;
